requirements.txt
django==1.11.10
djangorestframework==3.7.7
mysqlclient==1.3.12
boto3==1.5.35
django-storages==1.6.5
